third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/xhr-breakpoints-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests XHR breakpoints.


Running: testXHRBreakpoint
Script execution paused.
Call stack:
    0) sendRequest (xhr-breakpoints.js:14)
    1)  (:1)
Script execution resumed.

Running: testPauseOnAnyXHR
Script execution paused.
Script execution resumed.
Script execution paused.
Script execution resumed.

Running: testDisableBreakpoint
Script execution paused.
Script execution resumed.
